Some small refacto and function reordering
