--- demos/multigrid/Laplacian1D.cpp
@@ -1,45 +1,0 @@
-#pragma once
-#include "samurai_new/petsc_assembly.hpp"
-
-//-------------------//
-//     Laplacian     //
-// Implementation 1D //
-//-------------------//
-
-// Prediction (order 1)
-template<class Mesh>
-void assemble_prediction_impl(std::integral_constant<std::size_t, 1>, Mat& A, Mesh& mesh)
-{
-    using mesh_id_t = typename Mesh::mesh_id_t;
-
-    auto min_level = mesh[mesh_id_t::cells].min_level();
-    auto max_level = mesh[mesh_id_t::cells].max_level();
-    for(std::size_t level=min_level+1; level<=max_level; ++level)
-    {
-        auto set = samurai::intersection(mesh[mesh_id_t::cells_and_ghosts][level],
-                                        mesh[mesh_id_t::cells][level-1])
-                .on(level);
-
-        std::array<double, 3> pred{{1./8, 0, -1./8}};
-        set([&](const auto& i, const auto&)
-        {
-            for(int ii=i.start; ii<i.end; ++ii)
-            {
-                auto i_cell = static_cast<int>(mesh.get_index(level, ii));
-                MatSetValue(A, i_cell, i_cell, 1., INSERT_VALUES);
-
-                int sign_i = (ii & 1)? -1: 1;
-
-                for(int is = -1; is<2; ++is)
-                {
-                    auto i1 = static_cast<int>(mesh.get_index(level - 1, (ii>>1) + is));
-                    double v = -sign_i*pred[is + 1];
-                    MatSetValue(A, i_cell, i1, v, INSERT_VALUES);
-                }
-
-                auto i0 = static_cast<int>(mesh.get_index(level - 1, (ii>>1)));
-                MatSetValue(A, i_cell, i0, -1., INSERT_VALUES);
-            }
-        });
-    }
-}
